Removing JSON validation for data with custom variables
Summary:
{F684972231}
Currently, the view for configuring a rule does not allow an invalid json, even if it contains data elements that would be substituted.

Since we have no way of validating in the front end if a data element would be substituted by a valid json, This fix skips this validation when a data element is found.

Reviewed By: mrharel

Differential Revision: D32886833

fbshipit-source-id: c0ba4613bfe102bbb20681e02595e987cc50a867